docker_main/requirements.txt
fastapi
pydantic
Pillow
numpy
uvicorn
yolov5
ultralytics
huggingface-hub == 0.15.1 
python-multipart
